docs+cli: document new features, bump to 1.5.0, fix extract on Python 3.9
Documents the public features now on main and bumps the version (minor):
- Selector parser (#123): CSS/XPath, find_all, find_by_text, find_similar
- pyscrappy extract CLI (URL -> .md/.json/.txt/.html)
- ScraperConfig(impersonate=...) TLS fingerprint impersonation (#124)

README gains key-feature bullets, a Selector quick-start section, the
pyscrappy[stealth] extra, a command-line extract section, the impersonate
config option, and a stealth subsection under Proxies and blocked sites.
Version bumped to 1.5.0 across pyproject, __init__, and server.json.

Also fixes CI on Python 3.9: extract lived in mcp/agent.py, which imports
the MCP stack (fastmcp, 3.10+), so importing it broke 3.9. Moved
run_extract + the CLI entry to a new MCP-free pyscrappy/cli.py (entry
point pyscrappy.cli:main); chat still uses the agent, imported lazily.
