+++ src/pyscrappy/mcp/agent.py
@@ -118,70 +118,12 @@ async def run_agent(
     return stopped
 
 
-def run_extract(
-    url: str,
-    out_path: str,
-    css_selector: str | None = None,
-    render_js: bool = False,
-) -> str:
-    """Scrape ``url`` and write the result to ``out_path``. The output format is
-    chosen from the file extension:
-
-    - ``.md``   -> Markdown (``ScrapeResult.to_markdown``)
-    - ``.json`` -> JSON (``ScrapeResult.to_json``)
-    - ``.txt``  -> extracted page text
-    - ``.html`` -> raw page HTML
-
-    With ``--css-selector`` the matched elements' text is extracted instead of the
-    whole page. Returns a short status line for the CLI to print.
-    """
-    from pyscrappy import GenericScraper, scrape
-
-    ext = out_path.rsplit(".", 1)[-1].lower() if "." in out_path else ""
-    if ext not in {"json", "md", "txt", "html"}:
-        raise ValueError(f"unsupported output extension {ext!r}; use .md, .json, .txt, or .html")
-
-    if ext == "html":
-        # Raw HTML: the structured scrape result doesn't retain the source markup,
-        # so fetch the page's HTML directly.
-        with GenericScraper() as gs:
-            content = gs.http.get_html(url)
-    else:
-        selectors = {"match": css_selector} if css_selector else None
-        result = scrape(url, selectors=selectors, render_js=render_js)
-        if ext == "json":
-            content = result.to_json()
-        elif ext == "md":
-            content = result.to_markdown()
-        else:  # txt
-            if css_selector:
-                content = "\n".join(row.get("match", "") for row in result.data)
-            else:
-                content = "\n\n".join(item.get("text", {}).get("text", "") for item in result.data)
-
-    with open(out_path, "w", encoding="utf-8") as f:
-        f.write(content)
-    return f"wrote {len(content)} chars to {out_path}"
-
-
 def main() -> None:
     parser = argparse.ArgumentParser(
         prog="pyscrappy",
         description="Let a local LLM use PyScrappy's scrapers as tools (via Ollama).",
     )
     sub = parser.add_subparsers(dest="command", required=True)
-
-    extract = sub.add_parser(
-        "extract", help="Scrape a URL straight to a file (.md/.json/.txt/.html)."
-    )
-    extract.add_argument("url", help="The URL to scrape.")
-    extract.add_argument("output", help="Output file; format inferred from its extension.")
-    extract.add_argument(
-        "--css-selector", default=None, help="Extract only elements matching this CSS selector."
-    )
-    extract.add_argument(
-        "--render-js", action="store_true", help="Render JavaScript with a browser first."
-    )
 
     chat = sub.add_parser("chat", help="Ask a local model a question it can answer with scrapers.")
     chat.add_argument("prompt", help="What to ask, in plain language.")
@@ -196,16 +138,6 @@ def main() -> None:
     chat.add_argument("--json", action="store_true", help="Print the raw scraper result as JSON.")
 
     args = parser.parse_args()
-
-    if args.command == "extract":
-        status = run_extract(
-            args.url,
-            args.output,
-            css_selector=args.css_selector,
-            render_js=args.render_js,
-        )
-        print(status)
-        return
 
     if args.command == "chat":
         import anyio
